requirements.txt
cryptography
fastapi
fitz
pandas
pymongo
PyMuPDF
python-dotenv
requests
s3transfer
SQLAlchemy
urllib3
passlib
uvicorn
alembic
psycopg2-binary
sqlalchemy==2.0.36
python-jose==3.3.0
python-multipart